+++ src/main/java/com/dlnu/byname/domain/entity/UserRoleDO.java
@@ -30,25 +30,25 @@ public class UserRoleDO extends BaseDO {
     /**
      * 用户id
      */
-    private Integer userId;
+    private Long userId;
     /**
      * 角色id
      */
-    private Integer roleId;
+    private Long roleId;
 
-    public Integer getUserId() {
+    public Long getUserId() {
         return userId;
     }
 
-    public void setUserId(Integer userId) {
+    public void setUserId(Long userId) {
         this.userId = userId;
     }
 
-    public Integer getRoleId() {
+    public Long getRoleId() {
         return roleId;
     }
 
-    public void setRoleId(Integer roleId) {
+    public void setRoleId(Long roleId) {
         this.roleId = roleId;
     }
